+++ src/FolkerKinzel.Strings.Tests/ReadOnlySpanExtensionTests.cs
@@ -161,21 +161,21 @@ public class ReadOnlySpanExtensionTests
     //}
 
 
-    [DataTestMethod]
-    [DataRow("test", 0)]
-    [DataRow("", 0)]
-    [DataRow(" ", 1)]
-    [DataRow(" test", 1)]
-    public void GetTrimmedStartTest1(string input, int result)
-        => Assert.AreEqual(result, input.AsSpan().GetTrimmedStart());
+    //[DataTestMethod]
+    //[DataRow("test", 0)]
+    //[DataRow("", 0)]
+    //[DataRow(" ", 1)]
+    //[DataRow(" test", 1)]
+    //public void IndexOfAnyExceptTest1(string input, int result)
+    //    => Assert.AreEqual(result, input.AsSpan().IndexOfAnyExcept());
 
-    [DataTestMethod]
-    [DataRow("test", 4)]
-    [DataRow("", 0)]
-    [DataRow(" ", 0)]
-    [DataRow("test    ", 4)]
-    public void GetTrimmedLength1(string input, int length)
-        => Assert.AreEqual(length, input.AsSpan().GetTrimmedLength());
+    //[DataTestMethod]
+    //[DataRow("test", 4)]
+    //[DataRow("", 0)]
+    //[DataRow(" ", 0)]
+    //[DataRow("test    ", 4)]
+    //public void GetTrimmedLength1(string input, int length)
+    //    => Assert.AreEqual(length, input.AsSpan().GetTrimmedLength());
 
 
     [DataTestMethod]