+++ src/FolkerKinzel.Strings/md/ReleaseNotes.md
@@ -11,6 +11,7 @@ whole package.
 - **Breaking Change:** The .NET Framework 4.5 support has ended. A .NET Framework 4.6.1 DLL is part of the nuget package instead.
 - **Breaking Change:** Removed the obsolete method `StringBuilder.NormalizeNewLinesTo`.
 - **Breaking Change:** Removed the obsolete method `String.NormalizeNewLinesTo`.
+- **BreakingChange:** The `GetTrimmedStart` and `GetTrimmedLength` methods have been removed.
 - Deterministic build
 - Performance: Raised the maximum allowed stackalloc size for Char to 256 (according to the internal constant 
 `System.String.StackallockCharBufferSizeLimit`) from the .NET sources and for Byte to 512.
@@ -26,6 +27,18 @@ int GetTrimmedStart(this ReadOnlySpan<char>, char);
 int GetTrimmedStart(this Span<char>, char);
 int GetTrimmedLength(this ReadOnlySpan<char>, char);
 int GetTrimmedLength(this Span<char>, char);
+```
+&nbsp;
+- New Polyfills for the ReadOnlySpan&lt;Char&gt; struct (.NET 6.0, .NET 5.0, .NET Core 3.1, .NET Standard 2.1, .NET Standard 2.0, .NET Framework 4.6.1)
+```csharp
+int IndexOfAnyExcept(this ReadOnlySpan<char>, char);
+int LastIndexOfAnyExcept(this ReadOnlySpan<char>, char);
+```
+&nbsp;
+- New Polyfills for the Span&lt;Char&gt; struct (.NET 6.0, .NET 5.0, .NET Core 3.1, .NET Standard 2.1, .NET Standard 2.0, .NET Framework 4.6.1)
+```csharp
+int IndexOfAnyExcept(this Span<char>, char);
+int LastIndexOfAnyExcept(this Span<char>, char);
 ```
 &nbsp;
 - New Polyfills for the Span&lt;Char&gt; struct (.NET Standard 2.1, .NET Standard 2.0, .NET Framework 4.6.1):
@@ -47,6 +60,13 @@ string TrimStart(this string, string?);
 StringBuilder Trim(this StringBuilder, string?);
 StringBuilder TrimEnd(this StringBuilder, string?);
 StringBuilder TrimStart(this StringBuilder, string?);
+```
+&nbsp;
+- New Polyfills for the Memory&lt;Char&gt; struct (.NET Standard 2.1, .NET Standard 2.0, .NET Framework 4.6.1)
+```csharp
+Memory<char> Trim(this Memory<char>);
+Memory<char> TrimEnd(this Memory<char>);
+Memory<char> TrimStart(this Memory<char>);
 ```
 &nbsp;
 > **Project reference:** On some systems, the content of the CHM file in the Assets is blocked. Before opening the file right click on the file icon, select Properties, and **check the "Allow" checkbox** - if it is present - in the lower right corner of the General tab in the Properties dialog.